remove warnings

+++ RcppAD/inst/include/cppad/local/reverse.hpp
@@ -333,7 +333,7 @@ void ADFun<Base>::myReverse(size_t p, const VectorBase &w, size_t dep_var_index,
     tape_point tp;
     for(it=op_mark_index_.begin();it!=op_mark_index_.end();it++){
       tp=tp_[*it];
-      for(int i=0;i<CppAD::NumRes(tp.op);i++)
+      for(size_t i=0;i<CppAD::NumRes(tp.op);i++)
 	for(j = 0; j < p; j++)
 	  Partial[tp.var_index-i*p+j]=0;
     }
